requirements.txt
absl-py==2.1.0
acme==3.1.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.3.0
apache-beam==2.57.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==23.2.0
av==12.1.0
awscli==1.37.2
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
blis==1.2.0
boto3==1.36.2
botocore==1.36.2
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certbot==3.1.0
certifi==2023.11.17
cffi==1.16.0
cftime==1.6.3
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmake==3.29.2
cmdstanpy==1.2.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.5
ConfigArgParse==1.7
configobj==5.0.9
confluent-kafka==2.6.0
contourpy==1.2.0
copernicusmarine==2.0.0
crcmod==1.7
cryptography==44.0.0
cssselect2==0.7.0
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.10
dask==2025.1.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.1.1
distlib==0.3.8
distro==1.9.0
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
docker==7.1.0
docopt==0.6.2
docstring-parser==0.15
docutils==0.16
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.115.6
fastavro==1.9.5
fasteners==0.19
faster-whisper==1.0.2
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
Flask-SSLify==0.1.5
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozendict==2.4.2
frozenlist==1.4.1
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googlesearch-python==1.2.4
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5netcdf==1.5.0
h5py==3.11.0
hdfs==2.7.3
holidays==0.59
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.6
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==0.6.8
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
josepy==1.15.0
json5==0.9.22
jsonpickle==3.2.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kagglehub==0.3.6
keras==3.7.0
keyring==8.7
keyrings.alt==5.0.2
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd==1.2.0
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
logzero==1.7.0
lxml==5.2.1
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.13
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
murmurhash==1.0.12
mysqlclient==2.2.1
namex==0.0.8
narwhals==1.8.3
nbclient==0.9.0
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
netCDF4==1.6.5
networkx==3.4.2
nltk==3.9.1
notebook==7.1.1
notebook_shim==0.2.4
numcodecs==0.15.0
numpy==1.26.4
objsize==0.7.0
onnxruntime==1.18.0
openai==1.33.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.3
partd==1.4.2
patsy==0.5.6
pdfkit==1.0.0
peewee==3.17.3
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.2.0
pluggy==1.4.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
prophet==1.1.6
proto-plus==1.24.0
protobuf==5.29.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==1.4.2
pydyf==0.9.0
Pygments==2.17.2
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pymongo==4.8.0
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pyphen==0.14.0
pypiwin32==223
pyreadline3==3.4.1
pyRFC3339==2.0.1
pystac==1.12.1
pytest==8.1.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==2.0.7
python-multipart==0.0.20
python-socketio==5.12.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.8
referencing==0.33.0
regex==2024.7.24
reportlab==4.1.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.1
safetensors==0.5.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==71.1.0
shell_gpt==1.4.3
simple-websocket==1.1.0
six==1.16.0
smart-open==7.1.0
smartapi-python==1.4.8
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
spotify2ytmusic==0.9.30
sqlparse==0.2.4
srsly==2.5.1
stack-data==0.6.3
stanio==0.5.1
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.38.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.0
textblob==0.19.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.48.1
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
ultralytics==8.3.63
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
utils==1.0.2
uvicorn==0.34.0
virtualenv==20.26.2
voluptuous==0.15.2
waitress==3.0.2
wasabi==1.1.3
watchdog==4.0.2
wcwidth==0.2.13
weasel==0.4.1
weasyprint==61.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==14.2
Werkzeug==3.0.2
wheel==0.43.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xarray==2025.1.2
yarl==1.9.4
yfinance==0.2.38
yolo==0.3.1
ytmusicapi==1.8.2
zarr==2.18.4
zopfli==0.2.3
zstandard==0.23.0
